pyproject.toml
[project]
name = "google-calendar-mcp"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.10.1",
    "google-api-python-client>=2.175.0",
    "google-auth>=2.40.3",
    "google-auth-oauthlib>=1.2.2",
]
